% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfnormfns.R
\name{CDS.HN.fx.denom}
\alias{CDS.HN.fx.denom}
\title{Support function for \code{CDS.HN.lik}}
\usage{
CDS.HN.fx.denom(x, sigma1)
}
\arguments{
\item{x}{distances}

\item{sigma1}{half normal parameter}
}
\value{
vector of detection probabilities
}
\description{
Not to be called by user; helper function to CDS.HN.lik
}
