% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfnormfns.R
\name{CDS.HN.fx}
\alias{CDS.HN.fx}
\title{Support function for \code{CDS.HN.lik}}
\usage{
CDS.HN.fx(x, sigma1, w)
}
\arguments{
\item{x}{distances}

\item{sigma1}{half normal parameter}

\item{w}{trunction}
}
\value{
component of likelihood
}
\description{
Not to be called by user; helper function to CDS.HN.lik
}
