% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfnormfns.R
\name{CDS.HN.lik}
\alias{CDS.HN.lik}
\title{Likelihood of half-normal CDS detection function}
\usage{
CDS.HN.lik(par, distances, truncation)
}
\arguments{
\item{par}{parameter vector being optimised}

\item{distances}{detection distances}

\item{truncation}{truncation distance}
}
\value{
evaluated likelihood for HN CDS
}
\description{
To be used in conjunction with \code{optim()} to find MLEs of sigma parameter of half normal detection function
}
