% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazratefns.R
\name{CDS.HR.fx.denom}
\alias{CDS.HR.fx.denom}
\title{Support function for \code{CDS.HR.lik}}
\usage{
CDS.HR.fx.denom(x, sigma1, b)
}
\arguments{
\item{x}{detection distances}

\item{sigma1}{shape parameter of hazard rate}

\item{b}{scale parameter of hazard rate}
}
\value{
component of likelihood
}
\description{
Not to be called by user; helper function to CDS.HR.lik
}
