% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazratefns.R
\name{CDS.HR.fx}
\alias{CDS.HR.fx}
\title{Support function for \code{CDS.HR.lik}}
\usage{
CDS.HR.fx(x, sigma1, b, w)
}
\arguments{
\item{x}{distances}

\item{sigma1}{shape parameter of hazard rate}

\item{b}{scale parameter of hazard rate}

\item{w}{truncation distance}
}
\value{
component of likelihood
}
\description{
Not to be called by user; helper function to CDS.HR.lik
}
