% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazratefns.R
\name{CDS.HR.lik}
\alias{CDS.HR.lik}
\title{Likelihood of hazard rate CDS detection function}
\usage{
CDS.HR.lik(par, distances, truncation)
}
\arguments{
\item{par}{parameters of the hazard rate detection function}

\item{distances}{detection distances}

\item{truncation}{truncation distance}
}
\value{
evaluated likelihood for HR CDS
}
\description{
Likelihood of hazard rate CDS detection function
}
