% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfn.R
\name{detfn}
\alias{detfn}
\title{Distance sampling detection function}
\usage{
detfn(z, pars, key = "HN", adjn = 0, adjt = "cos", w = max(z))
}
\arguments{
\item{z}{generic distance (radial or perpendicular) - can be scalar or vector}

\item{pars}{the detection function parameter values (HN:s,adj1,adj2,...)(HR:s,b,adj1,adj2,...)(UNI:adj1,adj2,...)}

\item{key}{the detection function key, default is Half Normal (HR and UNI might be alternatives)}

\item{adjn}{if 0, no adjustment terms are used, if > 0, number of adjustments to use}

\item{adjt}{type of adjustments to use, defaults to cosine series}

\item{w}{truncation distance, only fundamental for adjustments, by default the max of the distances}
}
\value{
probability of detection
}
\description{
Computes probability of detection at a given distance when provided with key function, adjustments and parameters
}
