% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePhat.R
\name{estimate.Phat}
\alias{estimate.Phat}
\title{Utility function to estimate detection probability given joint likelihood parameter estimates}
\usage{
estimate.Phat(key, params, survey.truncation)
}
\arguments{
\item{key}{key function (HN or HR)}

\item{params}{estimated parameters from the combined likelihood with either HN or HR key}

\item{survey.truncation}{truncation distance for detection function}
}
\value{
Phat estimated probability of detection, integrated over detection distances out to w
}
\description{
Utility function to estimate detection probability given joint likelihood parameter estimates
}
