% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combolikHN.R
\name{full.lik.fx.HN}
\alias{full.lik.fx.HN}
\title{Combined likelihood with half normal detection and half normal gradient}
\usage{
full.lik.fx.HN(par, detectdists, gpsdists, truncdet, truncden)
}
\arguments{
\item{par}{parameters of combined likelihood (sigma of detection, sigma of gradient, beta gradient)}

\item{detectdists}{detection distances}

\item{gpsdists}{distances from feature from collared animals}

\item{truncdet}{truncation distance of distance sampling}

\item{truncden}{truncation distance of collaring study}
}
\value{
value of combined likelihood with HN detection function
}
\description{
Combined likelihood with half normal detection and half normal gradient
}
