% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combolikHR.R
\name{fx.denomHR}
\alias{fx.denomHR}
\title{Support function for full.lik.fx.HR}
\usage{
fx.denomHR(x, sigma1, b, sigma2, beta, w)
}
\arguments{
\item{x}{detection distances}

\item{sigma1}{shape parameter of hazard rate detection fn}

\item{b}{scale parameter of hazard rate detection fn}

\item{sigma2}{parameter of density gradient}

\item{beta}{attraction/repulsion parameter of density gradient}

\item{w}{truncation distance (distance sampling)}
}
\value{
component of likelihood
}
\description{
Not to be called by user; support function for \code{full.lik.fx.HR}
}
