% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combolikHN.R
\name{fx}
\alias{fx}
\title{Support function for full.lik.fx.HN}
\usage{
fx(x, sigma1, sigma2, beta, w)
}
\arguments{
\item{x}{detection distances}

\item{sigma1}{parameter of half normal detection fn}

\item{sigma2}{parameter of density gradient}

\item{beta}{attraction/repulsion parameter of density gradient}

\item{w}{truncation distance}
}
\value{
component of likelihood
}
\description{
Not to be called by user; support function for \code{full.lik.fx.HN}
}
