% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitylik.R
\name{likdensity}
\alias{likdensity}
\title{Likelihood function to estimate parameters of the density gradient}
\usage{
likdensity(par, distances, truncation)
}
\arguments{
\item{par}{parameters of density gradient}

\item{distances}{distances of animal distribution from transect}

\item{truncation}{truncation distance}
}
\value{
evaluated likelihood of density gradient function
}
\description{
Likelihood function to estimate parameters of the density gradient
}
