% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitylik.R
\name{pi.x.den}
\alias{pi.x.den}
\title{Component to modelling the density gradient}
\usage{
pi.x.den(x, sigma, beta)
}
\arguments{
\item{x}{the perpendicular (or radial) distance}

\item{sigma}{a gaussian standard deviation}

\item{beta}{a multiplicative constant, helping to improve the fit notice that if beta>0, animals avoid linear structure, if beta<0, they prefer it}
}
\value{
component of likelihood
}
\description{
#' Not to be called by user; support function for \code{likdensity}
}
