% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitylik.R
\name{pi.x}
\alias{pi.x}
\title{One form of modelling the density gradient}
\usage{
pi.x(x, sigma, beta, w)
}
\arguments{
\item{x}{the perpendicular distance}

\item{sigma}{a gaussian standard deviation}

\item{beta}{a multiplicative constant, helping to improve the fit notice that if beta>0, animals avoid linear structure, if beta<0, they prefer it}

\item{w}{truncation distance}
}
\value{
component of likelihood
}
\description{
Not to be called by user; support function for \code{likdensity}
}
