% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizebias.R
\name{size.bias}
\alias{size.bias}
\title{Compute size-bias adjusted group size estimates}
\usage{
size.bias(key, survey.data, survey, grad.data = real$dist,
  survey.truncation, gradient.max.distance, plotit = FALSE)
}
\arguments{
\item{key}{key function (HN or HR)}

\item{survey.data}{perpendicular detection distances from distance sampling survey}

\item{survey}{complete data frame of distance sampling survey data}

\item{grad.data}{animal telemetry data}

\item{survey.truncation}{truncation distance for distance sampling survey}

\item{gradient.max.distance}{maximum distance from transects for telemetry data}

\item{plotit}{flag determining plot of regression scatter plot}
}
\value{
adjusted group size
}
\description{
Compute size-bias adjusted group size estimates
}
