% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydesignsummary.R
\name{survey.design.summary}
\alias{survey.design.summary}
\title{Extract information about a distance sampling survey from survey data frame}
\usage{
survey.design.summary(survey, survey.truncation)
}
\arguments{
\item{survey}{data frame containing 'usual' distance sampling data}

\item{survey.truncation}{}
}
\value{
named list containing number of detections, survey effort and average group size
}
\description{
Extract information about a distance sampling survey from survey data frame
}
