% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbootstrap.R
\name{varbootstrap}
\alias{varbootstrap}
\title{Variance estimation via bootstrap over transects}
\usage{
varbootstrap(keyfn, survey, gpsdata, disttrunc = 150,
  gradientmaxdist = 100, Nboot, alpha = 0.05, plotit = TRUE,
  conversion = 10000)
}
\arguments{
\item{keyfn}{key function of detection function (HN or HR)}

\item{survey}{distance sampling survey data frame}

\item{gpsdata}{animal location data frame from telemetry}

\item{disttrunc}{truncation distance for distance sampling component}

\item{gradientmaxdist}{maximum distance from transect of telemetry data}

\item{Nboot}{number of bootstrap replicates}

\item{alpha}{Type I error rate for computing CI bounds}

\item{plotit}{logical indicating whether to produce histogram of bootstrap estimates}

\item{conversion}{conversion factor to get density units reported}
}
\value{
parameter estimates of all replicates and upper and lower confidence interval bounds on individual density estimates
}
\description{
Variance estimation via bootstrap over transects
}
